rename ebscoEdsSearch to EbscoArticleSearch
