+++ lib/controller/ebsco/articleSearch.js
@@ -1,5 +1,5 @@
 import searchParser from '../../services/searchParser';
-import ebscoEdsSearch from '../../services/ebscoEdsSearch';
+import ebscoArticleSearch from '../../services/ebscoArticleSearch';
 
 export const articleSearch = function* articleSearch(domainName) {
     const { userId, password, profile } = this.domain;
@@ -9,7 +9,7 @@ export const articleSearch = function* articleSearch(domainName) {
         this.status = 401;
         return;
     }
-    const result = yield ebscoEdsSearch(this.query, sessionToken, authToken);
+    const result = yield ebscoArticleSearch(this.query, sessionToken, authToken);
 
     this.body = searchParser(result);
 };
@@ -18,5 +18,5 @@ export const articleSearchPure = function* articleSearchPure(domainName) {
     const { userId, password, profile } = this.domain;
     const authToken = yield this.getAuthenticationToken(domainName, userId, password);
     const sessionToken = yield this.getSessionToken(domainName, profile, authToken);
-    this.body = yield ebscoEdsSearch(this.query, sessionToken, authToken);
+    this.body = yield ebscoArticleSearch(this.query, sessionToken, authToken);
 };
