+++ lib/services/ebscoArticleSearch.js
